+++ src/Pim/Bundle/EnrichBundle/Resources/public/js/manager/product-manager.js
@@ -17,22 +17,15 @@ define([
     ) {
         return {
             productValues: null,
-            productPromises: {},
             get: function (id) {
-                if (!(id in this.productPromises)) {
-                    this.productPromises[id] = $.getJSON(Routing.generate('pim_enrich_product_rest_get', { id: id }))
-                        .then(_.bind(function (product) {
-                            return this.generateMissing(product);
-                        }, this))
-                        .then(function (product) {
-                            mediator.trigger('product:action:post_fetch', product);
+                return $.getJSON(Routing.generate('pim_enrich_product_rest_get', { id: id }))
+                    .then(this.generateMissing)
+                    .then(function (product) {
+                        mediator.trigger('pim_enrich:form:product:action:post_fetch', product);
 
-                            return product;
-                        })
-                        .promise();
-                }
-
-                return this.productPromises[id];
+                        return product;
+                    })
+                    .promise();
             },
             save: function (id, data) {
                 return $.ajax({
@@ -40,10 +33,10 @@ define([
                     url: Routing.generate('pim_enrich_product_rest_get', {id: id}),
                     contentType: 'application/json',
                     data: JSON.stringify(data)
-                }).then(_.bind(function (data) {
-                    this.productPromises[id] = $.Deferred().resolve(data).promise();
+                }).then(_.bind(function (product) {
+                    mediator.trigger('pim_enrich:form:entity:action:post_save', product);
 
-                    return data;
+                    return product;
                 }, this));
             },
             remove: function (id) {
@@ -52,14 +45,7 @@ define([
                     url: Routing.generate('pim_enrich_product_remove', {id: id}),
                     headers: { accept: 'application/json' },
                     data: { _method: 'DELETE' }
-                }).then(_.bind(function () {
-                    delete this.productPromises[id];
-                }, this));
-            },
-            clear: function (id) {
-                if (id in this.productPromises) {
-                    delete this.productPromises[id];
-                }
+                });
             },
             getValues: function (product) {
                 return AttributeManager.getAttributesForProduct(product).then(function (attributes) {