+++ src/Pim/Bundle/TransformBundle/Normalizer/Flat/LocaleNormalizer.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Pim\Bundle\TransformBundle\Normalizer\Flat;
+
+use Pim\Bundle\CatalogBundle\Model\LocaleInterface;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
+/**
+ * Normalize a locale
+ *
+ * @author    Sanchez Julien <julien@akeneo.com>
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class LocaleNormalizer implements NormalizerInterface
+{
+    /** @var array */
+    protected $supportedFormats = ['csv', 'flat'];
+
+    /**
+     * {@inheritdoc}
+     */
+    public function normalize($locale, $format = null, array $context = [])
+    {
+        return ['code' => $locale->getCode()];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supportsNormalization($data, $format = null)
+    {
+        return $data instanceof LocaleInterface && in_array($format, $this->supportedFormats);
+    }
+}